Fix App Attest security gaps and split proxy into modules
- Validate challenge belongs to requesting device (was ignoring device ID)
- Store and verify rpIdHash during assertions to prevent cross-app key reuse
- Split index.js (972 LOC) into index.js, attest.js, crypto.js (all <500 LOC)
- Consolidate duplicate derSignatureToRaw into single parameterized function
- Clarify KV namespace placeholder in wrangler.toml
